Update README.md
Add the cider equivalent for your emacs code for running each form one at a time in the repl. Thanks for this, been too lazy to write my own :) I've upgraded to CIDER and had to dig a bit to find the equivalent statements. Thought I'd share in case anyone else needs it.